Fix #1001173, #1001174 - Correctly expose abstract val/var to JDT
Abstract val/var were being ignored in the ``ScalaStructureBuilder`` and
therefore no ``IJavaElement`` was being created. Indeed, abstract val/var are
represented in scalac as ``DefTree``, and the attached symbol has the ACCESSOR
flag set (that because the symbol that is attached to abstract val/var is
always the getter symbol, i.e., its accessor).

The fix consisted in simply not pruning ``DefTree`` that have the ACCESSOR flag
set (this fixed #1001173).

A nice side effect of the fix is that missing implementation of abstract val/var
are now reported in Java sources (which fixes #1001174).