Correctly index referenced to var's assignments (i.e., calls to the setter)
When looking for reference of a field assignment, both calls to the var's
setter (i.e., assignment) and getter should be returned. For this to be
working, the ``ScalaIndexBuilder`` is now checking if the name of a ``RefTree``
is a setter name, and if that is so a reference to the getter name (which is
the actual field name) is instead added in the indexer.

An additional interesting point is the element selection in the editor. When in
the editor the user right click on an element to find his references, the
symbol attached to the selected code element is typecheck, and in the case of
an var assignement the typechecked symbol is the setter symbol, while for the
very same untyped (i.e., not typechecked) tree node the attached symbol is the
getter symbol.

In ``ScalaCompilationUnit.codeSelect`` setter symbols are turned into getter
ones because the ``ScalaIndexBuilder`` does not indexes the setter name, but
only the getter name (which is the actual var's name), as explained in the
first paragraph of this commit.