More abstract dealings with Scala compilation units.
We need a way to treat uniformly different kinds of Scala-based sources. Until now
we routed everything through `ScalaCompilationUnit`, that extends the JDT `CompilationUnit`
class. This tight connection is counter-productive, and in most cases, unnecessary.

This commit is a step towards extracting a set of interfaces for other Scala-based
sources to benefit of reconciliation ('errors-as-you-type'), hyperlinking and completion
engine that already exists.

* extracted `InteractiveCompilationUnit` as a super type of `ScalaCompilationUnit`
* switched from `ScalaCompilationUnit` to `InteractiveCompilationUnit` wherever possible, to allow
hyperlinking to work in the scala-worksheet project
* added a utility method to retrieve the compilation unit associated with an open editor that
uses the platform adapter protocol to allow other plugins to participate.