Improved the product configuration
It depends directly on the root pom, to avoid the attach source problem
The repositories have been added to the product pom
It uses the same properties/profile setup as the other poms
Added update sites, and remove log errors
Create a product to be able to start on the Scala perspective
Customize the product to show Scala IDE information and branding
Set default preferences to not display the 'new Scala IDE version' dialog, but set the
recommanded preferences.
Remove the source bundles from the product

Update the initialization of default preference value in code, to be able to use
an external file with different default values.