Add test cases for semantic highlighting
There are some symbols that can not yet highlighted due to too
less information provided by the compiler. Once the compiler
stores more information in the AST the test cases should be
activated and implemented (if necessary).

The test case TypeTest#set_is_a_type is not needed any more, the
problem seems to be fixed.