No timeout when running tests.
Added an environment option to suppress timeouts when running tests. This will make
tests more repeatable. The option is passed by default in the maven build for tests,
but if you want to run the IDE in that mode (for instance, in the debugger), make sure
to add -Dsdtcore.notimeouts to your command line options.