Fix #1001233 - Graceful termination of debug sessions
Major refactoring of actor code in debugger

Extracted a common superclass (``BaseDebuggerActor``) for all actors defined in
the debugger module. This avoids duplication and makes sure that actors
terminates consistently. This base class for actors provides a number of
interesting functionalities:

* Automatic logging of all messages processed by the actor. This can be quite
useful when debugging. However, this functionality is currently commented
because it generates a lot of noise. Before enabling it, we need to add a TRACE
log level in the logger (I'll open a ticket for this).

* Stackable behaviors. By calling `become/unbecome` an actor can easily
modify the shape of messages that it processes.

* Cleaner actor's life-cylce. Basically, when starting an actor, `postStart`
hook is guaranteed to be called before the processing of any message. Then, the
actor will process messages accordingly to its `behavior` implementation.  If
an `Exit` or `PoisonPill` message is processed by `this` actor, the actor is
guaranteed to execute `preExit` before terminating.

Last, I have also added a number of FIXMEs for some problematic code that could
cause the debug session to freeze. I'll be opening tickets for each one of
these issues, so that we don't forget to fix them by the final V2.1 release
(Re #1001307, Re #1001308, Re #1001310, Re #1001311).