Switch from 'body' to `loadedType` in mark occurrences.
`body` (in fact, `askType`) will trigger a new type-checking pass regardless of where inside the type-checker you happen to be. This is most likely wrong, the we need to remove this option from the presentation compiler API. In the meantime, switched to using `waitLoadedTyped`, that is simply waiting until the current type checking phase has finished.