Fix #1001330 - Race condition in ``ScalaDeclarationHyperlinkComputer``
It turns out that calling ``Tree.toString`` outside of the Presentation
Compiler Thread can trigger lazy evaluation of Tree's members, which is a
problem since the compiler's data structure are not thread-safe.

I'm hoping that ``Tree.shortClass`` will provide enough information when the
list of hyperlinks is empty (if you have a better candidate, please say
something, or open a pull request).