Add semantic highlighting for symbols
This adds:
- Semantic highlighting support + test case for Symbols
- Symbol color option in the "Syntax Coloring" preferences

Scala symbols do not have their own representation by a Symbol
type stored in the AST. Thus, they are highlighted as methods because
internally they are only passed as the apply method of class scala.Symbol.

To avoid highlighting symbols as methods scalariform is used to parse
them correctly and to make it possible to give them their own color in
the preferences.

Fixes #1001364