Fixed ``TcpipSpy`` on Mac
It's pretty weird, but it looks like when the main TcpipSpy thread terminates, the
spawned ``in/outThread`` are also killed. This is extremely surprising, because
the ``TcpipSpy`` threads are not daemons!

Anyway, a couple of calls to ``join`` on the main Thread do the trick and now I
can spy on a remote debugging sessions. This is really useful!