Fixed #1001304 - NSDHO when hovering the Scala launch shortcut
When checking if a runnable JUnit test class is available in the compilation
unit containing the selected code element, it is sometime possible to get a
Scala compiler exception: ``No symbol does not have an owner``. This happens
when trying to access the ``owner`` member of a ``Symbol`` that is of type
``NoSymbol``.

I'm unfortunately not sure on the whys this can happen, but the net result is
particularly bad. The pragmatic solution is to check that the ``symbol`` is a
``ClassSymbol``, and simply bail out when it isn't.

No test is available because we honestly have no clue about how to reproduce
this.