Fixed the build for Juno.
The build was broken because Juno changed an interface method from taking a raw type to using a proper type, like this:

connect(Map args, ..)

to

connect(Map<String, String> args, ..)

We used to implement this method by using an existential type, but overriding rules prevent it in the new case. So we resort to Java, where we bridge the method in question towards one that has the proper generic type signature in both Indigo and Juno.

Fixed #1001374.