Fix #1001380 - Show Run As > JUnit Test when right clicking on a source that contains runnable JUnit Test classes
If a source is closed (i.e., it hasn't been typechecked yet - this is
important!) and, through the package explorer, you right click and select Run
As, no "JUnit Test" option used to be shown.

The problem was that if the source is not opened (and typechecked), the
``@Test`` annotation can only be found in the AST, and not in the symbol. The
fix is to traverse the AST and check for ``@Test`` occurrences.

There is only one occurrence when this doesn't work, which is when a class
declares no tests but it inheriths tests from a superclass. The problem here is
that we cannot access the AST of the superclass, if the superclass is defined
in a different compilation unit (there is no way to go from a symbol to its
AST). The right way to fix this is to us the ScalaSearch component, once we
will have it. I've opened another ticket for keeping track of this (Re

As part of this commit I've also made some improvements to the existing testing
class (formerly named ``LaunchableTesters``, now splitted into two distinct
classes: ``MainMethodFinderTest`` and ``RunAsTest``).

The work this ticket (Re #1001380) was somewhat motivated by (or a consequence
of) Re #1001304.

Finally, I'm quite convinced that the ``ScalaLaunchShortcut.getMainMethods``
could use some similar love, but this should be done in a separate PR.