Added tests for Re #1001380
There is currently no way to tell if a class, that declares no JUnit tests, it
inheriths some JUnit tests from a superclass. The issue is that we cannot
access the AST of the superclass if the superclass is defined in a different
compilation unit (there is no way to go from a symbol to its AST). The right
way to fix this is to us the ScalaSearch component, once we will have it.

Make sure to enable the tests that are part of this commit once Re #1001380 is
fixed.