Implemented JUnit test finder logic in Scala.
Works for both Scala and Java sources, and adds a 'Scala JUnit' test kind and shortcut for 'Run As'.

We hook this test finder using an internal extension point defined by `org.eclipse.jdt.junit.core`, `internal_testkinds`. This class is used both when right-clicking on a Java element and choosing "Run As - Scala JUnit test" and when hitting the "Search" button in the JUnit Run configuration dialog.

Alternatives were considered, but not pursued:
  * AspectJ to hook into the JUnit runner. Discarded as hackish, not future-proof and the risk of breaking the plain JDT plugin
  * implement a Scala JUnit configuration type, but that would duplicate more of the JUnit configuration type and involved rewriting dialogs and buttons

Fixed #1001275, Fixed #1000782.