``sbt.precompiled.version`` has to match ``scala.version``
It is really important that when we build the IDE, the sbt artifacts are
compiled against the exact same Scala compiler that is shipped in the IDE. To
this end, I've updated the value of ``sbt.precompiled.version`` to always match
the ``scala.version`` one. This is needed because while the Scala library is
ensured to be forward binary compatible across minor releases, no such
guarantee is provided for the Scala compiler (and the Sbt compiler-interface,
which we use in the IDE, does need to be linked against the Scala compiler).

As part of this commit, I've also made another small change to the property
``scala.major.minor.version`` and renamed it ``scala.era.major.version``, as
this what this property actually contains.