Add character literal completion
As for string literals, \ signs are automatically escaped to prevent
compiler errors.

Furthermore it turned out that the current partitioning strategy doesn't
work perfectly together with this component. Two consecutive apostrophes
aren't handled as their own partition since they are not valid Scala
syntax. Thus, the behavior that should only applied to existing
character literals is placed in the same component and must be moved if
the partitioning strategy for character literals ever changes.

Fix #1001403