Correct auto indent behavior in comments
The old behavior did not correctly indent text inside of comments
whereas the comments themselves are indented correctly. This commit
corrects the behavior by indent a new line to the same depth as the
previous line.

Another feature introduced is, that if the enter key is pressed while
the cursor is adjusted in an line not starting with a *, a completely
empty line is inserted instead of auto inserting a * sign at the begin
of the related line.

Furthermore, ScaladocAutoEditStrategy is renamed to
CommentAutoEditStrategy because this component does its auto edits not
only to Scaladoc but also to multi-line comments.