Correct syntax highlighting for while-keyword
While syntax highlighting correctly recognizes the while-keyword as a
keyword, semantic highlighting treats it as a method. Because semantic
highlighting is activated after its syntax counterpart, it overwrites
the previous detected colors.

The new behavior omits all occurrences of LabelDef in the AST because
they don't regularly appear in Scala sources but represent while- and
do-while-keywords internally in the compiler. Hence they must not be
highlighted.

Fix #1000984