Correctly list test methods in the JUnit search dialog
Fix the search button in JUnit test configuration tab. We need to use the same
`id` in the test runner extension as the JUnit4 runner, otherwise the button
will look for JUnit *3* tests.

Since all the code is in private methods in UI classes, there is no test for this.
I tried it out and works well for Java and Scala classes, with the Java and
Scala test runners (so using their ID does not seem to impact anything other than
our own code).

Fixed #1001474.