Properly initialize Scala editor (quick fix/interactive error reporting works even before a save).
Compilation units in the JDT are backed up by a `buffer`, holding the source that 
hasn't been saved yet. We use the buffer to get notification of character changes,
and in turn notify the presentation compiler. Without a proper buffer, errors (and
quick fixes) are not reported either.

Buffers are created lazily by the JDT, and we mimic the JDT code in 
`buildStructure`.

Fixed #1001094, Fixed #1001337.