Correct regression error in ScalaCodeScanner
After refactoring the ScalaCodeScanner I discarded all white space
token because I thought it is an optimization (they must never
be colored). The offset error now occurres because of another
optimization that is done in a component built-in in Eclipse:
org.eclipse.jface.text.rules.DefaultDamagerRepairer. Whenever this
component detects style ranges of the same type that are consecutive,
only the first style range is applied and then expanded for all
following ranges. Because white space token are omitted it can be that
a color range is expanded although it is not allowed.

To correct the behavior white space token are not discarded anymore.

Fix #1001481